+++ Tests/IntegrationTests.Shared/ObjectIntegrationTests.cs
@@ -276,64 +276,4 @@ namespace IntegrationTests
 
     }  // ObjectIntegrationTests
 
-    [TestFixture, Preserve(AllMembers = true)]
-    public class RealmMigrationTests
-    {
-        [Test]
-        public void TriggerMigrationBySchemaVersion()
-        {
-            // Arrange
-            var config1 = new RealmConfiguration("ChangingVersion.realm");
-            Realm.DeleteRealm(config1);  // ensure start clean
-            var realm1 = Realm.GetInstance(config1);
-            // new database doesn't push back a version number
-            Assert.That(config1.SchemaVersion, Is.EqualTo(RealmConfiguration.NotVersioned));
-            realm1.Close();
-
-            // Act
-            var config2 = config1.ConfigWithPath("ChangingVersion.realm");
-            config2.SchemaVersion = 99;
-            Realm realm2 = null;  // should be updated by DoesNotThrow
-
-            // Assert
-            Assert.DoesNotThrow( () => realm2 = Realm.GetInstance(config2) ); // same path, different version, should auto-migrate quietly
-            Assert.That(realm2.Config.SchemaVersion, Is.EqualTo(99));
-            realm2.Close();
-
-        }
-
-        [Test]
-        [Ignore("this tests for a condition that's only available under manual migrations, which we lack")]
-        public void TriggerMigrationBySchemaEditing()
-        {
-            
-            // NOTE to regnerate the bundled database go edit the schema in Person.cs and comment/uncomment ExtraToTriggerMigration
-            // running in between and saving a copy with the added field
-            // this should never be needed as this test just needs the Realm to need migrating
-            TestHelpers.CopyBundledDatabaseToDocuments(
-                "ForMigrationsToCopyAndMigrate.realm", "NeedsMigrating.realm");
-
-            // Assert
-            Realm realm1 = null;
-            Assert.Throws<RealmMigrationNeededException>( () => realm1 = Realm.GetInstance("NeedsMigrating.realm") );
-        }
-
-        [Test]
-        public void MigrationTriggersDelete()
-        {
-            // Arrange
-            var config = new RealmConfiguration("MigrateWWillRecreate.realm", true);
-            Realm.DeleteRealm(config);
-            Assert.False(File.Exists(config.DatabasePath));
-
-            TestHelpers.CopyBundledDatabaseToDocuments(
-                "ForMigrationsToCopyAndMigrate.realm", "MigrateWWillRecreate.realm");
-
-            // Act - should cope by deleting and silently recreating
-            var realm = Realm.GetInstance(config);
-
-            // Assert
-            Assert.That(File.Exists(config.DatabasePath));
-        }
-    }
 }
